requirements.txt
anthropic==0.94.1
httpx==0.28.1
google-genai==1.73.0
openai==2.31.0
telegram-markdown-entities==0.1.1
